@@ -15,7 +15,7 @@
   ],
   "license": "BSD-3-Clause",
   "require": {
-    "php": ">=7.4.0",
+    "php": ">=8.0.0",
     "ext-mbstring": "*",
     "ext-json": "*"
   },
